+++ src/databricks/azext_databricks/aaz/latest/databricks/access_connector/_list.py
@@ -25,12 +25,14 @@ class List(AAZCommand):
     """
 
     _aaz_info = {
-        "version": "2022-10-01-preview",
+        "version": "2024-05-01",
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/providers/microsoft.databricks/accessconnectors", "2022-10-01-preview"],
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.databricks/accessconnectors", "2022-10-01-preview"],
+            ["mgmt-plane", "/subscriptions/{}/providers/microsoft.databricks/accessconnectors", "2024-05-01"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.databricks/accessconnectors", "2024-05-01"],
         ]
     }
+
+    AZ_SUPPORT_PAGINATION = True
 
     def _handler(self, command_args):
         super()._handler(command_args)
@@ -117,7 +119,7 @@ class List(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2022-10-01-preview",
+                    "api-version", "2024-05-01",
                     required=True,
                 ),
             }
@@ -213,6 +215,13 @@ class List(AAZCommand):
                 serialized_name="provisioningState",
                 flags={"read_only": True},
             )
+            properties.refered_by = AAZListType(
+                serialized_name="referedBy",
+                flags={"read_only": True},
+            )
+
+            refered_by = cls._schema_on_200.value.Element.properties.refered_by
+            refered_by.Element = AAZStrType()
 
             system_data = cls._schema_on_200.value.Element.system_data
             system_data.created_at = AAZStrType(
@@ -279,7 +288,7 @@ class List(AAZCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2022-10-01-preview",
+                    "api-version", "2024-05-01",
                     required=True,
                 ),
             }
@@ -375,6 +384,13 @@ class List(AAZCommand):
                 serialized_name="provisioningState",
                 flags={"read_only": True},
             )
+            properties.refered_by = AAZListType(
+                serialized_name="referedBy",
+                flags={"read_only": True},
+            )
+
+            refered_by = cls._schema_on_200.value.Element.properties.refered_by
+            refered_by.Element = AAZStrType()
 
             system_data = cls._schema_on_200.value.Element.system_data
             system_data.created_at = AAZStrType(