+++ src/databricks/azext_databricks/aaz/latest/databricks/workspace/_wait.py
@@ -20,7 +20,7 @@ class Wait(AAZWaitCommand):
 
     _aaz_info = {
         "resources": [
-            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.databricks/workspaces/{}", "2023-02-01"],
+            ["mgmt-plane", "/subscriptions/{}/resourcegroups/{}/providers/microsoft.databricks/workspaces/{}", "2024-05-01"],
         ]
     }
 
@@ -120,7 +120,7 @@ class Wait(AAZWaitCommand):
         def query_parameters(self):
             parameters = {
                 **self.serialize_query_param(
-                    "api-version", "2023-02-01",
+                    "api-version", "2024-05-01",
                     required=True,
                 ),
             }
@@ -176,6 +176,9 @@ class Wait(AAZWaitCommand):
             )
 
             properties = cls._schema_on_200.properties
+            properties.access_connector = AAZObjectType(
+                serialized_name="accessConnector",
+            )
             properties.authorizations = AAZListType()
             properties.created_by = AAZObjectType(
                 serialized_name="createdBy",
@@ -185,11 +188,24 @@ class Wait(AAZWaitCommand):
                 serialized_name="createdDateTime",
                 flags={"read_only": True},
             )
+            properties.default_catalog = AAZObjectType(
+                serialized_name="defaultCatalog",
+            )
+            properties.default_storage_firewall = AAZStrType(
+                serialized_name="defaultStorageFirewall",
+            )
             properties.disk_encryption_set_id = AAZStrType(
                 serialized_name="diskEncryptionSetId",
                 flags={"read_only": True},
             )
             properties.encryption = AAZObjectType()
+            properties.enhanced_security_compliance = AAZObjectType(
+                serialized_name="enhancedSecurityCompliance",
+            )
+            properties.is_uc_enabled = AAZBoolType(
+                serialized_name="isUcEnabled",
+                flags={"read_only": True},
+            )
             properties.managed_disk_identity = AAZObjectType(
                 serialized_name="managedDiskIdentity",
             )
@@ -233,6 +249,18 @@ class Wait(AAZWaitCommand):
                 flags={"read_only": True},
             )
 
+            access_connector = cls._schema_on_200.properties.access_connector
+            access_connector.id = AAZStrType(
+                flags={"required": True},
+            )
+            access_connector.identity_type = AAZStrType(
+                serialized_name="identityType",
+                flags={"required": True},
+            )
+            access_connector.user_assigned_identity_id = AAZStrType(
+                serialized_name="userAssignedIdentityId",
+            )
+
             authorizations = cls._schema_on_200.properties.authorizations
             authorizations.Element = AAZObjectType()
 
@@ -244,6 +272,14 @@ class Wait(AAZWaitCommand):
             _element.role_definition_id = AAZStrType(
                 serialized_name="roleDefinitionId",
                 flags={"required": True},
+            )
+
+            default_catalog = cls._schema_on_200.properties.default_catalog
+            default_catalog.initial_name = AAZStrType(
+                serialized_name="initialName",
+            )
+            default_catalog.initial_type = AAZStrType(
+                serialized_name="initialType",
             )
 
             encryption = cls._schema_on_200.properties.encryption
@@ -309,6 +345,32 @@ class Wait(AAZWaitCommand):
                 flags={"required": True},
             )
 
+            enhanced_security_compliance = cls._schema_on_200.properties.enhanced_security_compliance
+            enhanced_security_compliance.automatic_cluster_update = AAZObjectType(
+                serialized_name="automaticClusterUpdate",
+            )
+            enhanced_security_compliance.compliance_security_profile = AAZObjectType(
+                serialized_name="complianceSecurityProfile",
+            )
+            enhanced_security_compliance.enhanced_security_monitoring = AAZObjectType(
+                serialized_name="enhancedSecurityMonitoring",
+            )
+
+            automatic_cluster_update = cls._schema_on_200.properties.enhanced_security_compliance.automatic_cluster_update
+            automatic_cluster_update.value = AAZStrType()
+
+            compliance_security_profile = cls._schema_on_200.properties.enhanced_security_compliance.compliance_security_profile
+            compliance_security_profile.compliance_standards = AAZListType(
+                serialized_name="complianceStandards",
+            )
+            compliance_security_profile.value = AAZStrType()
+
+            compliance_standards = cls._schema_on_200.properties.enhanced_security_compliance.compliance_security_profile.compliance_standards
+            compliance_standards.Element = AAZStrType()
+
+            enhanced_security_monitoring = cls._schema_on_200.properties.enhanced_security_compliance.enhanced_security_monitoring
+            enhanced_security_monitoring.value = AAZStrType()
+
             parameters = cls._schema_on_200.properties.parameters
             parameters.aml_workspace_id = AAZObjectType(
                 serialized_name="amlWorkspaceId",
@@ -329,7 +391,6 @@ class Wait(AAZWaitCommand):
             parameters.enable_no_public_ip = AAZObjectType(
                 serialized_name="enableNoPublicIp",
             )
-            _WaitHelper._build_schema_workspace_custom_boolean_parameter_read(parameters.enable_no_public_ip)
             parameters.encryption = AAZObjectType()
             parameters.load_balancer_backend_pool_name = AAZObjectType(
                 serialized_name="loadBalancerBackendPoolName",
@@ -357,6 +418,7 @@ class Wait(AAZWaitCommand):
             _WaitHelper._build_schema_workspace_custom_boolean_parameter_read(parameters.require_infrastructure_encryption)
             parameters.resource_tags = AAZObjectType(
                 serialized_name="resourceTags",
+                flags={"read_only": True},
             )
             parameters.storage_account_name = AAZObjectType(
                 serialized_name="storageAccountName",
@@ -370,6 +432,14 @@ class Wait(AAZWaitCommand):
                 serialized_name="vnetAddressPrefix",
             )
             _WaitHelper._build_schema_workspace_custom_string_parameter_read(parameters.vnet_address_prefix)
+
+            enable_no_public_ip = cls._schema_on_200.properties.parameters.enable_no_public_ip
+            enable_no_public_ip.type = AAZStrType(
+                flags={"read_only": True},
+            )
+            enable_no_public_ip.value = AAZBoolType(
+                flags={"required": True},
+            )
 
             encryption = cls._schema_on_200.properties.parameters.encryption
             encryption.type = AAZStrType(
@@ -390,6 +460,9 @@ class Wait(AAZWaitCommand):
             resource_tags = cls._schema_on_200.properties.parameters.resource_tags
             resource_tags.type = AAZStrType(
                 flags={"read_only": True},
+            )
+            resource_tags.value = AAZFreeFormDictType(
+                flags={"required": True},
             )
 
             private_endpoint_connections = cls._schema_on_200.properties.private_endpoint_connections